+++ lib/ahoy/engine.rb
@@ -1,6 +1,8 @@
 module Ahoy
   class Engine < ::Rails::Engine
     initializer "ahoy", after: "sprockets.environment" do
+      Ahoy.logger ||= Rails.logger
+
       # allow Devise to be loaded after Ahoy
       require "ahoy/warden" if defined?(Warden)