+++ lib/ahoy/visit_properties.rb
@@ -1,7 +1,6 @@
-require "browser"
+require "cgi"
 require "device_detector"
-require "referer-parser"
-require "user_agent_parser"
+require "uri"
 
 module Ahoy
   class VisitProperties
@@ -21,23 +20,26 @@ module Ahoy
     private
 
     def utm_properties
-      landing_uri = Addressable::URI.parse(landing_page) rescue nil
-      landing_params = (landing_uri && landing_uri.query_values) || {}
+      landing_params = {}
+      begin
+        landing_uri = URI.parse(landing_page)
+        # could also use Rack::Utils.parse_nested_query
+        landing_params = CGI.parse(landing_uri.query) if landing_uri
+      rescue
+        # do nothing
+      end
 
       props = {}
       %w(utm_source utm_medium utm_term utm_content utm_campaign).each do |name|
-        props[name.to_sym] = params[name] || landing_params[name]
+        props[name.to_sym] = params[name] || landing_params[name].try(:first)
       end
       props
     end
 
     def traffic_properties
-      # cache for performance
-      @@referrer_parser ||= RefererParser::Parser.new
-
+      uri = URI.parse(referrer) rescue nil
       {
-        referring_domain: (Addressable::URI.parse(referrer).host.first(255) rescue nil),
-        search_keyword: (@@referrer_parser.parse(@referrer)[:term].first(255) rescue nil).presence
+        referring_domain: uri.try(:host).try(:first, 255)
       }
     end
 
@@ -60,6 +62,9 @@ module Ahoy
           device_type: device_type
         }
       else
+        raise "Add browser to your Gemfile to use legacy user agent parsing" unless defined?(Browser)
+        raise "Add user_agent_parser to your Gemfile to use legacy user agent parsing" unless defined?(UserAgentParser)
+
         # cache for performance
         @@user_agent_parser ||= UserAgentParser::Parser.new